backend/requirements.txt
# Core
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=0.19.0,<0.20.0

# Database
sqlalchemy>=1.4.23,<2.0.0
alembic>=1.7.3,<2.0.0
psycopg2-binary>=2.9.1,<3.0.0  # For PostgreSQL

# Data validation
pydantic>=1.8.0,<2.0.0
email-validator>=1.1.3,<2.0.0

# Utilities
tenacity>=8.0.1,<9.0.0  # For retrying database connections
python-dateutil>=2.8.2,<3.0.0
httpx>=0.19.0,<0.20.0  # For async HTTP requests
sse-starlette>=1.6.1  # For server-sent events streaming
psutil>=5.9.0,<6.0.0  # For system monitoring

# AI/LLM Integration
langchain>=0.0.300
ollama>=0.1.7
requests>=2.28.0

# Testing
pytest>=6.2.5,<7.0.0
pytest-cov>=2.12.1,<3.0.0
pytest-asyncio>=0.15.1,<0.16.0

# Code Quality
black==21.9b0
isort==5.9.3
flake8==3.9.2
mypy==0.910

# Testing
pytest==6.2.5
pytest-cov==2.12.1
pytest-asyncio==0.15.1
pytest-env==0.6.2

# Documentation
mkdocs==1.2.3
mkdocs-material==7.3.6
